feat: pd split test files
