+++ core/pipeline-driver/tests/utils.js
@@ -1,0 +1,7 @@
+
+const delay = d => new Promise(r => setTimeout(r, d));
+
+module.exports = {
+    delay
+};
+
